[gui] change zoom to floating point

+++ tracking/tracking_window.h
@@ -94,7 +94,7 @@ public:
     void report(QString string);
     void move_slice_to(tipl::vector<3,float> pos);
     bool can_map_to_mni(void);
-    void set_roi_zoom(int zoom);
+    void set_roi_zoom(float zoom);
     bool command(QString cmd,QString param = "",QString param2 = "");
 public slots:
     void restore_3D_window();
@@ -184,7 +184,7 @@ private slots:
     void on_actionMark_Tracts_on_T1W_T2W_triggered();
     void on_actionApply_Operation_triggered();
     void on_actionSave_Slices_to_DICOM_triggered();
-    void on_zoom_valueChanged(int arg1);
+    void on_zoom_valueChanged(double arg1);
 };
 
 #endif // TRACKING_WINDOW_H
